pyproject.toml
[project]
name = "gust"
version = "0.1.0"
description = "What's the main goal of this package?"
authors = [
    { name = "David Zhang" },
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.13"
dependencies = [
    "click==8.3.1",
    "google-genai==1.56.0",
    "markdown==3.8.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit==4.3.0",
    "pytest==8.4.1",
    "python-dotenv==1.2.1",
]

build = [
    "build==1.3.0",
]

[project.scripts]
gust = "gust.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
